Log file archiving
Each time the firmware is started, old log file for the same casu on the beaglebone is moved to ~/firmware/old-log. If when all of the casu-fw are run on the bbg, old logs are archived to .tar in ~/firmware/archives. New naming convention, to be used with bbg that have RTC enabled. Otherwise expect archive overwritting.
Setup instructions updated with info on .dtb.